+++ src/modal/DialogBoxProvider.jsx
@@ -14,6 +14,9 @@
  * 2025/12/01   1.0.2     ITA       Reverted back to default import of prop-types and tried another workaround for the vite issue.
  *                                  When importing this compoent into a vite based project from its published package the vite
  *                                  issue was still there.
+ * 2025/12/10   1.0.3     ITA       Changed the export of DialogBoxProvider from default to named.
+ *                                  This is to improve tree-shaking for consumers of the dialogboxjs package that use modern bundlers such as vite and WebPack, webpack.
+ *                                  useDialogBox function to throw if used within a component not wrapped inside a DialogBoxProvider.
  *
  */
 /** File: ./src/modal/DialogBoxProvider.jsx */
@@ -24,7 +27,7 @@ import PropTypes from 'prop-types';
 const DialogBoxContext = createContext();
 
 // Provider component
-export default function DialogBoxProvider({ children }) {
+export function DialogBoxProvider({ children }) {
     const funcs = useRef({});
 
     // Functions to manage modal state
@@ -58,4 +61,11 @@ DialogBoxProvider.propTypes = {
 };
 
 // Custom hook for convenience
-export const useDialogBox = () => useContext(DialogBoxContext);
+export const useDialogBox = () => {
+    const context = useContext(DialogBoxContext);
+
+    if (!context) { // undefined
+        throw new Error("Components using useDialogBox context must be wrapped within a DialogBoxProvider!");        
+    }
+    return context;
+};